+++ R/prediction_hurdle.R
@@ -1,0 +1,40 @@
+#' @rdname prediction
+#' @export
+prediction.hurdle <- 
+function(model, 
+         data = find_data(model, parent.frame()), 
+         at = NULL, 
+         type = c("response", "count", "prob", "zero"), 
+         ...) {
+    
+    type <- match.arg(type)
+    
+    # extract predicted values
+    data <- data
+    if (missing(data) || is.null(data)) {
+        pred <- predict(model, type = type, ...)
+        pred <- data.frame(fitted = pred[["fit"]])
+    } else {
+        # setup data
+        out <- build_datalist(data, at = at)
+        for (i in seq_along(out)) {
+            tmp <- predict(model, 
+                           newdata = out[[i]], 
+                           type = type, 
+                           se.fit = TRUE,
+                           ...)
+            out[[i]] <- cbind(out[[i]], fitted = tmp)
+            rm(tmp)
+        }
+        pred <- do.call("rbind", out)
+    }
+    pred[["se.fitted"]] <- NA_real_
+    
+    # obs-x-(ncol(data)+2) data frame
+    structure(pred, 
+              class = c("prediction", "data.frame"), 
+              row.names = seq_len(nrow(pred)),
+              at = if (is.null(at)) at else names(at), 
+              model.class = class(model),
+              type = type)
+}