+++ R/prediction_zeroinfl.R
@@ -1,0 +1,3 @@
+#' @rdname prediction
+#' @export
+prediction.zeroinfl <- prediction.hurdle