+++ NEWS.md
@@ -1,3 +1,8 @@
+# CHANGES TO prediction 0.1.17
+
+* Added `prediction.zeroinfl()` method for "zeroinfl" objects from **pscl**. (#1)
+* Added `prediction.hurdle()` method for "hurdle" objects from **pscl**. (#1)
+
 # CHANGES TO prediction 0.1.16
 
 * Added `prediction.plm()` method for "plm" objects from **plm**. (#1)